Move tabular WC code to kart/tabular
