+++ kart/repo.py
@@ -227,7 +227,7 @@ class KartRepo(pygit2.Repository):
         repo_root_path = repo_root_path.resolve()
         cls._ensure_exists_and_empty(repo_root_path)
         if not bare:
-            from kart.working_copy.base import BaseWorkingCopy
+            from kart.tabular.working_copy.base import BaseWorkingCopy
 
             BaseWorkingCopy.check_valid_creation_location(
                 wc_location, PotentialRepo(repo_root_path)
@@ -291,7 +291,7 @@ class KartRepo(pygit2.Repository):
         repo_root_path = repo_root_path.resolve()
         cls._ensure_exists_and_empty(repo_root_path)
         if not bare:
-            from kart.working_copy.base import BaseWorkingCopy
+            from kart.tabular.working_copy.base import BaseWorkingCopy
 
             BaseWorkingCopy.check_valid_creation_location(
                 wc_location, PotentialRepo(repo_root_path)
@@ -431,7 +431,7 @@ class KartRepo(pygit2.Repository):
         # Force writing to reflogs:
         self.config["core.logAllRefUpdates"] = "always"
         # Write working copy config:
-        from kart.working_copy.base import BaseWorkingCopy
+        from kart.tabular.working_copy.base import BaseWorkingCopy
 
         BaseWorkingCopy.write_config(self, wc_location, bare)
         if spatial_filter_spec:
@@ -655,7 +655,7 @@ class KartRepo(pygit2.Repository):
         allow_invalid_state=False,
         allow_unconnectable=False,
     ):
-        from kart.working_copy.base import BaseWorkingCopy
+        from kart.tabular.working_copy.base import BaseWorkingCopy
 
         return BaseWorkingCopy.get(
             self,
